moved data to resources folder
